+++ src/clients/AdtExecutor.ts
@@ -1,5 +1,10 @@
 import type { IAbapConnection, ILogger } from '@mcp-abap-adt/interfaces';
-import { ClassExecutor, type IClassExecutor } from '../executors';
+import {
+  ClassExecutor,
+  type IClassExecutor,
+  type IProgramExecutor,
+  ProgramExecutor,
+} from '../executors';
 
 export class AdtExecutor {
   private readonly connection: IAbapConnection;
@@ -12,5 +17,9 @@ export class AdtExecutor {
 
   getClassExecutor(): IClassExecutor {
     return new ClassExecutor(this.connection, this.logger);
+  }
+
+  getProgramExecutor(): IProgramExecutor {
+    return new ProgramExecutor(this.connection, this.logger);
   }
 }